Split out Status::FromErrno into status_posix.cc
