+++ Firestore/core/src/firebase/firestore/util/status_apple.mm
@@ -1,5 +1,5 @@
 /*
- * Copyright 2015, 2018 Google
+ * Copyright 2018 Google
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,8 +16,6 @@
 
 #include "Firestore/core/src/firebase/firestore/util/status.h"
 
-#include <cerrno>
-
 #include "Firestore/core/src/firebase/firestore/util/string_format.h"
 
 namespace firebase {
@@ -25,6 +23,10 @@ namespace firestore {
 namespace util {
 
 Status Status::FromNSError(NSError* error) {
+  if (!error) {
+    return Status::OK();
+  }
+
   NSError* original = error;
 
   while (error) {
